Fix bug when newline char occurs on buffer switch
In case the setOffset function was called with a value that is larger
than the internal size of the buffer, the behavior was undefined,
because the internal buffer was not correctly moved forward.

This didn't lead to a segfault, but was visible on a large files, where
it happened that a newline char was exactly at the offset position that
was out of bounds of the internal buffer. Because of this the lines
pointer was not correctly moved forward.